Moved the ListPagination component into the components folder
